update phase 1 tests
