+++ starter-code/1-chess-game/passoff/chess/extracredit/CastlingTests.java
@@ -2,10 +2,10 @@ package passoff.chess.extracredit;
 
 import chess.*;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import passoff.TestFactory;
+
+import static passoff.TestFactory.loadBoard;
 
 /**
  * Tests if the ChessGame implementation can handle Castling moves
@@ -17,270 +17,171 @@ import passoff.TestFactory;
  */
 public class CastlingTests {
 
-    private ChessBoard board;
-    private ChessPosition position;
-    private ChessPiece king;
-
-    @BeforeEach
-    public void setup() {
-        board = TestFactory.getNewBoard();
-    }
-
-
     @Test
     @DisplayName("White Team Castle")
     public void castleWhite() {
-
-        /*
-        | | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		|R| | | |K| | |R|
-         */
-
-        //set up king
-        king = TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.KING);
-        position = TestFactory.getNewPosition(1, 5);
-        board.addPiece(position, king);
-
-        //add both rooks
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(1, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-
-        //set up game
-        ChessGame game = TestFactory.getNewGame();
+        ChessBoard board = loadBoard("""
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | |K| | |R|
+                """);
+        ChessGame game = new ChessGame();
         game.setBoard(board);
         game.setTeamTurn(ChessGame.TeamColor.WHITE);
 
         //check that with nothing in way, king can castle
-        ChessMove queenSide = TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 3), null);
-        ChessMove kingSide = TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 7), null);
+        ChessPosition kingPosition = new ChessPosition(1, 5);
+        ChessMove queenSide = new ChessMove(kingPosition, new ChessPosition(1, 3), null);
+        ChessMove kingSide = new ChessMove(kingPosition, new ChessPosition(1, 7), null);
 
-        Assertions.assertTrue(game.validMoves(position).contains(queenSide),
+        Assertions.assertTrue(game.validMoves(kingPosition).contains(queenSide),
                 "ChessGame validMoves did not contain valid queen-side castle move");
-        Assertions.assertTrue(game.validMoves(position).contains(kingSide),
+        Assertions.assertTrue(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves did not contain valid king-side castle move");
 
         //queen side castle works correctly
         Assertions.assertDoesNotThrow(() -> game.makeMove(queenSide));
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(1, 5)),
-                "After castling move, a piece is still present in the king's initial position");
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(1, 1)),
-                "After castling move, a piece is still present in the rook's initial position");
-
-        ChessPiece foundKing = game.getBoard().getPiece(TestFactory.getNewPosition(1, 3));
-        Assertions.assertNotNull(foundKing, "After castling move, no piece found at king's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.KING, foundKing.getPieceType(),
-                "Found piece at king's position is not a king");
-        Assertions.assertEquals(ChessGame.TeamColor.WHITE, foundKing.getTeamColor(),
-                "Found piece at king's position is the wrong team color");
-
-        ChessPiece foundRook = game.getBoard().getPiece(TestFactory.getNewPosition(1, 4));
-        Assertions.assertNotNull(foundRook, "After castling move, no piece found at rook's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.ROOK, foundRook.getPieceType(),
-                "Found piece at rook's position is not a rook");
-        Assertions.assertEquals(ChessGame.TeamColor.WHITE, foundRook.getTeamColor(),
-                "Found piece at rook's position is the wrong team color");
-
+        Assertions.assertEquals(loadBoard("""
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | |K|R| | | |R|
+                """), game.getBoard());
 
         //reset board
-        board = TestFactory.getNewBoard();
-        board.addPiece(TestFactory.getNewPosition(1, 5),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.KING));
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(1, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
+        board = loadBoard("""
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | |K| | |R|
+                """);
         game.setBoard(board);
         game.setTeamTurn(ChessGame.TeamColor.WHITE);
 
         //king side castle works correctly
         Assertions.assertDoesNotThrow(() -> game.makeMove(kingSide));
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(1, 5)),
-                "After castling move, a piece is still present in the king's initial position");
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(1, 8)),
-                "After castling move, a piece is still present in the rook's initial position");
-
-        foundKing = game.getBoard().getPiece(TestFactory.getNewPosition(1, 7));
-        Assertions.assertNotNull(foundKing, "After castling move, no piece found at king's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.KING, foundKing.getPieceType(),
-                "Found piece at king's position is not a king");
-        Assertions.assertEquals(ChessGame.TeamColor.WHITE, foundKing.getTeamColor(),
-                "Found piece at king's position is the wrong team color");
-
-        foundRook = game.getBoard().getPiece(TestFactory.getNewPosition(1, 6));
-        Assertions.assertNotNull(foundRook, "After castling move, no piece found at rook's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.ROOK, foundRook.getPieceType(),
-                "Found piece at rook's position is not a rook");
-        Assertions.assertEquals(ChessGame.TeamColor.WHITE, foundRook.getTeamColor(),
-                "Found piece at rook's position is the wrong team color");
+        Assertions.assertEquals(loadBoard("""
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | | |R|K| |
+                """), game.getBoard());
     }
 
 
     @Test
     @DisplayName("Black Team Castle")
     public void castleBlack() {
-
-        /*
-        |r| | | |k| | |r|
-		| |p| | | | | |q|
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		|R| | | | | | | |
-         */
-
-        //set up king
-        king = TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.KING);
-        position = TestFactory.getNewPosition(8, 5);
-        board.addPiece(position, king);
-
-        //set up rooks
-        board.addPiece(TestFactory.getNewPosition(8, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(8, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-
-        //extra decoy pieces
-        board.addPiece(TestFactory.getNewPosition(7, 2),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.PAWN));
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(7, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.QUEEN));
-
-        //set up game
-        ChessGame game = TestFactory.getNewGame();
+        ChessBoard board = loadBoard("""
+                |r| | | |k| | |r|
+                | |p| | | | | |q|
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | | | | | |
+                """);
+        ChessGame game = new ChessGame();
         game.setBoard(board);
         game.setTeamTurn(ChessGame.TeamColor.BLACK);
 
         //check that with nothing in way, king can castle
-        ChessMove queenSide = TestFactory.getNewMove(TestFactory.getNewPosition(8, 5),
-                TestFactory.getNewPosition(8, 3), null);
-        ChessMove kingSide = TestFactory.getNewMove(TestFactory.getNewPosition(8, 5),
-                TestFactory.getNewPosition(8, 7), null);
+        ChessPosition kingPosition = new ChessPosition(8, 5);
+        ChessMove queenSide = new ChessMove(kingPosition, new ChessPosition(8, 3), null);
+        ChessMove kingSide = new ChessMove(kingPosition, new ChessPosition(8, 7), null);
 
-        Assertions.assertTrue(game.validMoves(position).contains(queenSide),
+        Assertions.assertTrue(game.validMoves(kingPosition).contains(queenSide),
                 "ChessGame validMoves did not contain valid queen-side castle move");
-        Assertions.assertTrue(game.validMoves(position).contains(kingSide),
+        Assertions.assertTrue(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves did not contain valid king-side castle move");
 
         //queen side castle works correctly
         Assertions.assertDoesNotThrow(() -> game.makeMove(queenSide));
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(8, 5)),
-                "After castling move, a piece is still present in the king's initial position");
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(8, 1)),
-                "After castling move, a piece is still present in the rook's initial position");
-
-        ChessPiece foundKing = game.getBoard().getPiece(TestFactory.getNewPosition(8, 3));
-        Assertions.assertNotNull(foundKing, "After castling move, no piece found at king's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.KING, foundKing.getPieceType(),
-                "Found piece at king's position is not a king");
-        Assertions.assertEquals(ChessGame.TeamColor.BLACK, foundKing.getTeamColor(),
-                "Found piece at king's position is the wrong team color");
-
-        ChessPiece foundRook = game.getBoard().getPiece(TestFactory.getNewPosition(8, 4));
-        Assertions.assertNotNull(foundRook, "After castling move, no piece found at rook's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.ROOK, foundRook.getPieceType(),
-                "Found piece at rook's position is not a rook");
-        Assertions.assertEquals(ChessGame.TeamColor.BLACK, foundRook.getTeamColor(),
-                "Found piece at rook's position is the wrong team color");
+        Assertions.assertEquals(loadBoard("""
+                | | |k|r| | | |r|
+                | |p| | | | | |q|
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | | | | | |
+                """), game.getBoard());
 
 
         //reset board
-        board = TestFactory.getNewBoard();
-        board.addPiece(TestFactory.getNewPosition(8, 5),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.KING));
-        board.addPiece(TestFactory.getNewPosition(8, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(8, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(7, 2),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.PAWN));
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(7, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.QUEEN));
+        board = loadBoard("""
+                |r| | | |k| | |r|
+                | |p| | | | | |q|
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | | | | | |
+                """);
         game.setBoard(board);
         game.setTeamTurn(ChessGame.TeamColor.BLACK);
 
         //king side castle works correctly
         Assertions.assertDoesNotThrow(() -> game.makeMove(kingSide));
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(8, 5)),
-                "After castling move, a piece is still present in the king's initial position");
-        Assertions.assertNull(game.getBoard().getPiece(TestFactory.getNewPosition(8, 8)),
-                "After castling move, a piece is still present in the rook's initial position");
-
-        foundKing = game.getBoard().getPiece(TestFactory.getNewPosition(8, 7));
-        Assertions.assertNotNull(foundKing, "After castling move, no piece found at king's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.KING, foundKing.getPieceType(),
-                "Found piece at king's position is not a king");
-        Assertions.assertEquals(ChessGame.TeamColor.BLACK, foundKing.getTeamColor(),
-                "Found piece at king's position is the wrong team color");
-
-        foundRook = game.getBoard().getPiece(TestFactory.getNewPosition(8, 6));
-        Assertions.assertNotNull(foundRook, "After castling move, no piece found at rook's new position");
-        Assertions.assertEquals(ChessPiece.PieceType.ROOK, foundRook.getPieceType(),
-                "Found piece at rook's position is not a rook");
-        Assertions.assertEquals(ChessGame.TeamColor.BLACK, foundRook.getTeamColor(),
-                "Found piece at rook's position is the wrong team color");
+        Assertions.assertEquals(loadBoard("""
+                |r| | | | |r|k| |
+                | |p| | | | | |q|
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | | | | | |
+                """), game.getBoard());
     }
 
 
     @Test
     @DisplayName("Cannot Castle Through Pieces")
     public void castlingBlockedByTeam() {
-        /*
-        | | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		|R| |B| |K| |Q|R|
-         */
-
-        //set up king
-        king = TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.KING);
-        position = TestFactory.getNewPosition(1, 5);
-        board.addPiece(position, king);
-
-        //set up rooks
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(1, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-
-        //own team pieces in way
-        board.addPiece(TestFactory.getNewPosition(1, 3),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.BISHOP));
-        board.addPiece(TestFactory.getNewPosition(1, 7),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.QUEEN));
-
-        //set up game
-        ChessGame game = TestFactory.getNewGame();
+        ChessBoard board = loadBoard("""
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| |B| |K| |Q|R|
+                """);
+        ChessGame game = new ChessGame();
         game.setBoard(board);
+        game.setTeamTurn(ChessGame.TeamColor.WHITE);
+
+        //check that with nothing in way, king can castle
+        ChessPosition kingPosition = new ChessPosition(1, 5);
+        ChessMove queenSide = new ChessMove(kingPosition, new ChessPosition(1, 3), null);
+        ChessMove kingSide = new ChessMove(kingPosition, new ChessPosition(1, 7), null);
 
         //make sure king cannot castle
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 3), null)),
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(queenSide),
                 "ChessGame validMoves contained an invalid castling move");
-
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 7), null)),
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves contained an invalid castling move");
     }
 
@@ -288,54 +189,26 @@ public class CastlingTests {
     @Test
     @DisplayName("Cannot Castle in Check")
     public void castlingBlockedByEnemy() {
-
-        /*
-        |r| | |B|k| | |r|
-		| | | | | | | | |
-		| | | | | |R| | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| |K| | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-         */
-
-        //set up king
-        king = TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.KING);
-        position = TestFactory.getNewPosition(8, 5);
-        board.addPiece(position, king);
-
-        //set up rooks
-        board.addPiece(TestFactory.getNewPosition(8, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(8, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.ROOK));
-
-        //enemy piece in way
-        board.addPiece(TestFactory.getNewPosition(8, 4),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.BISHOP));
-
-        //enemy threatening in between position
-        board.addPiece(TestFactory.getNewPosition(6, 6),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-
-        //decoy enemy king
-        board.addPiece(TestFactory.getNewPosition(3, 2),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.KING));
-
-        //set up game
-        ChessGame game = TestFactory.getNewGame();
+        ChessBoard board = loadBoard("""
+                |r| | |B|k| | |r|
+                | | | | | | | | |
+                | | | | | |R| | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | |K| | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                """);
+        ChessGame game = new ChessGame();
         game.setBoard(board);
 
         //make sure king cannot castle on either side
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(8, 5),
-                        TestFactory.getNewPosition(8, 3), null)),
+        ChessPosition kingPosition = new ChessPosition(8, 5);
+        ChessMove queenSide = new ChessMove(kingPosition, new ChessPosition(8, 3), null);
+        ChessMove kingSide = new ChessMove(kingPosition, new ChessPosition(8, 7), null);
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(queenSide),
                 "ChessGame validMoves contained an invalid castling move");
-
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(8, 5),
-                        TestFactory.getNewPosition(8, 7), null)),
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves contained an invalid castling move");
     }
 
@@ -343,57 +216,26 @@ public class CastlingTests {
     @Test
     @DisplayName("Cannot Castle After Moving")
     public void noCastleAfterMove() throws InvalidMoveException {
-
-        /*
-        | | | | | | | | |
-		|p| | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		|R| | | |K| | |R|
-         */
-
-        //set up king
-        king = TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.KING);
-        position = TestFactory.getNewPosition(1, 5);
-        board.addPiece(position, king);
-
-        //add both rooks
-        board.addPiece(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-        board.addPiece(TestFactory.getNewPosition(1, 8),
-                TestFactory.getNewPiece(ChessGame.TeamColor.WHITE, ChessPiece.PieceType.ROOK));
-
-        //enemy pawn for filler moves
-        board.addPiece(TestFactory.getNewPosition(7, 1),
-                TestFactory.getNewPiece(ChessGame.TeamColor.BLACK, ChessPiece.PieceType.PAWN));
-
-        //set up game
-        ChessGame game = TestFactory.getNewGame();
+        ChessBoard board = loadBoard("""
+                | | | | | | | | |
+                |p| | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                | | | | | | | | |
+                |R| | | |K| | |R|
+                """);
+        ChessGame game = new ChessGame();
         game.setBoard(board);
         game.setTeamTurn(ChessGame.TeamColor.WHITE);
 
         //move left rook
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(1, 1),
-                TestFactory.getNewPosition(1, 4), null));
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(7, 1),
-                TestFactory.getNewPosition(6, 1), null));
-        /*
-        | | | | | | | | |
-		| | | | | | | | |
-		|p| | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | | | | | | |
-		| | | |R|K| | |R|
-         */
+        game.makeMove(new ChessMove(new ChessPosition(1, 1), new ChessPosition(1, 4), null));
+        game.makeMove(new ChessMove(new ChessPosition(7, 1), new ChessPosition(6, 1), null));
 
         //move rook back to starting spot
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(1, 4),
-                TestFactory.getNewPosition(1, 1), null));
+        game.makeMove(new ChessMove(new ChessPosition(1, 4), new ChessPosition(1, 1), null));
         /*
         | | | | | | | | |
 		| | | | | | | | |
@@ -405,21 +247,19 @@ public class CastlingTests {
 		|R| | | |K| | |R|
          */
 
-        //make sure king can't castle towards moved rook, but still can to unmoved rook
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 3), null)),
-                "ChessGame validMoves contained an invalid castling move");
+        ChessPosition kingPosition = new ChessPosition(1, 5);
+        ChessMove queenSide = new ChessMove(kingPosition, new ChessPosition(1, 3), null);
+        ChessMove kingSide = new ChessMove(kingPosition, new ChessPosition(1, 7), null);
 
-        Assertions.assertTrue(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 7), null)),
+        //make sure king can't castle towards moved rook, but still can to unmoved rook
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(queenSide),
+                "ChessGame validMoves contained an invalid castling move");
+        Assertions.assertTrue(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves did not contain valid king-side castle move");
 
         //move king
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(6, 1),
-                TestFactory.getNewPosition(5, 1), null));
-        game.makeMove(TestFactory.getNewMove(position, TestFactory.getNewPosition(1, 6), null));
+        game.makeMove(new ChessMove(new ChessPosition(6, 1), new ChessPosition(5, 1), null));
+        game.makeMove(new ChessMove(kingPosition, new ChessPosition(1, 6), null));
         /*
         | | | | | | | | |
 		| | | | | | | | |
@@ -432,9 +272,8 @@ public class CastlingTests {
          */
 
         //move king back to starting position
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(5, 1),
-                TestFactory.getNewPosition(4, 1), null));
-        game.makeMove(TestFactory.getNewMove(TestFactory.getNewPosition(1, 6), position, null));
+        game.makeMove(new ChessMove(new ChessPosition(5, 1), new ChessPosition(4, 1), null));
+        game.makeMove(new ChessMove(new ChessPosition(1, 6), kingPosition, null));
         /*
         | | | | | | | | |
 		| | | | | | | | |
@@ -447,14 +286,9 @@ public class CastlingTests {
          */
 
         //make sure king can't castle anymore
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 3), null)),
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(queenSide),
                 "ChessGame validMoves contained an invalid castling move");
-
-        Assertions.assertFalse(game.validMoves(position).contains(
-                TestFactory.getNewMove(TestFactory.getNewPosition(1, 5),
-                        TestFactory.getNewPosition(1, 7), null)),
+        Assertions.assertFalse(game.validMoves(kingPosition).contains(kingSide),
                 "ChessGame validMoves contained an invalid castling move");
     }
